Midterm Project
The goal of my code was to create a puzzle game where the user can drag
and drop individual puzzle pieces to form the completed picture.  My
code is not pretty and has a lot of redundant code that I’m sure could
be cleaned up with the use of an array, maybe. Unfortunately, there is
something wrong with my code as not all puzzle pieces are registering
as clicked and other pieces are all being clicked at once.  The problem
could come from my if statements but I’m not entirely sure.  Also,
strongly only puzzle pieces 1, 2 and 3 are working as intended and the
rest do not.  Furthermore, the dimensions of the pieces themselves are
not being properly defined as clicking just outside of a piece may be
registered as clicked (but not all sides ).  I was able to learn a lot
from this project and will admit that there was many frustrations to be
had when things weren’t working as planned, but it was also very
rewarding when I finally figured out what was wrong or missing.